Fix syntax highlighting in dark mode
This is actually caused by SSR. Previous code works perfectly fine in dev mode.
However, during SSR, accessing localStorage throws an error, which causes the code block to be stuck at light mode.
The solution is to give up rendering code block during SSR. This is intended to be a temporary solution since the next version of docusaurus will come with native support for switching between light and dark mode for code block.